requirements.txt
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
ta==0.11.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
MetaTrader5==5.0.45
